move person details to correct folder
